update controllers to include jwt

+++ app/controllers/contacto_controller.js
@@ -3,22 +3,29 @@ const Contacto = db.contacto;
 
 exports.create = (req, res) => {
   try {
-    const contacto = new Contacto({
-      nombre: req.body.nombre,
-      apellido: req.body.apellido,
-      empresa: req.body.empresa,
-      cargo: req.body.cargo,
-      email: req.body.email,
-      region: req.body.region,
-      ciudad: req.body.ciudad,
-      interes: req.body.interes,
-      canalPreferido: req.body.canalPreferido,
-      canales: req.body.canales,
-    });
-    contacto
-      .save()
-      .then(data => res.send(data))
-      .catch(error => console.log(error));
+    const validacion = chequearToken(
+      req.headers["authorization"].split(" ")[1]
+    );
+    if (validacion.resultado === "Autorizado") {
+      const contacto = new Contacto({
+        nombre: req.body.nombre,
+        apellido: req.body.apellido,
+        empresa: req.body.empresa,
+        cargo: req.body.cargo,
+        email: req.body.email,
+        region: req.body.region,
+        ciudad: req.body.ciudad,
+        interes: req.body.interes,
+        canalPreferido: req.body.canalPreferido,
+        canales: req.body.canales,
+      });
+      contacto
+        .save()
+        .then(data => res.send(data))
+        .catch(error => console.log(error));
+    } else {
+      res.status(401).send("Token inválido");
+    }
   } catch {
     res
       .status(400)
@@ -30,9 +37,16 @@ exports.create = (req, res) => {
 
 exports.findAll = (req, res) => {
   try {
-    Contacto.find()
-      .then(data => res.send(data))
-      .catch(error => console.log(error));
+    const validacion = chequearToken(
+      req.headers["authorization"].split(" ")[1]
+    );
+    if (validacion.resultado === "Autorizado") {
+      Contacto.find()
+        .then(data => res.send(data))
+        .catch(error => console.log(error));
+    } else {
+      res.status(401).send("Token inválido");
+    }
   } catch {
     res.status(400).send("Error");
   }
@@ -40,11 +54,22 @@ exports.findAll = (req, res) => {
 
 exports.findOne = (req, res) => {
   try {
-    Contacto.find({
-      $or: [{nombre: req.body.nombre}, {apellido: req.body.apellido}, {empresa: req.body.empresa}],
-    })
-      .then(data => res.send(data))
-      .catch(error => console.log(error));
+    const validacion = chequearToken(
+      req.headers["authorization"].split(" ")[1]
+    );
+    if (validacion.resultado === "Autorizado") {
+      Contacto.find({
+        $or: [
+          {nombre: req.body.nombre},
+          {apellido: req.body.apellido},
+          {empresa: req.body.empresa},
+        ],
+      })
+        .then(data => res.send(data))
+        .catch(error => console.log(error));
+    } else {
+      res.status(401).send("Token inválido");
+    }
   } catch {
     res.status(400).send("Error");
   }
@@ -52,17 +77,24 @@ exports.findOne = (req, res) => {
 
 exports.update = (req, res) => {
   try {
-    Contacto.updateOne(
-      {_id: req.params.id},
-      {
-        $set: {
-          nombre: req.body.nombre,
-          apellido: req.body.apellido,
-        },
-      }
-    )
-      .then(data => res.send(data))
-      .catch(error => console.log(error));
+    const validacion = chequearToken(
+      req.headers["authorization"].split(" ")[1]
+    );
+    if (validacion.resultado === "Autorizado") {
+      Contacto.updateOne(
+        {_id: req.params.id},
+        {
+          $set: {
+            nombre: req.body.nombre,
+            apellido: req.body.apellido,
+          },
+        }
+      )
+        .then(data => res.send(data))
+        .catch(error => console.log(error));
+    } else {
+      res.status(401).send("Token inválido");
+    }
   } catch {
     res.status(400).send("Error");
   }
@@ -70,9 +102,16 @@ exports.update = (req, res) => {
 
 exports.delete = (req, res) => {
   try {
-    Contacto.deleteOne({_id: req.params.id})
-      .then(data => res.send(data))
-      .catch(error => console.log(error));
+    const validacion = chequearToken(
+      req.headers["authorization"].split(" ")[1]
+    );
+    if (validacion.resultado === "Autorizado") {
+      Contacto.deleteOne({_id: req.params.id})
+        .then(data => res.send(data))
+        .catch(error => console.log(error));
+    } else {
+      res.status(401).send("Token inválido");
+    }
   } catch {
     res.status(400).send("Error");
   }
@@ -80,9 +119,16 @@ exports.delete = (req, res) => {
 
 exports.deleteAll = (req, res) => {
   try {
-    Contacto.deleteMany()
-      .then(data => res.send(data))
-      .catch(error => console.log(error));
+    const validacion = chequearToken(
+      req.headers["authorization"].split(" ")[1]
+    );
+    if (validacion.resultado === "Autorizado") {
+      Contacto.deleteMany()
+        .then(data => res.send(data))
+        .catch(error => console.log(error));
+    } else {
+      res.status(401).send("Token inválido");
+    }
   } catch {
     res.status(400).send("Error");
   }
